remove base test class
